% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plugin.R
\name{check_availability}
\alias{check_availability}
\title{Check if a context variable is available at a given node}
\usage{
check_availability(cv, node_name, node_order)
}
\description{
Check if a context variable is available at a given node
}
\keyword{internal}
