% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{classify_functions}
\alias{classify_functions}
\title{Classify exported functions by role}
\usage{
classify_functions(func_names, ns, package)
}
\description{
Classify exported functions by role
}
\keyword{internal}
