% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{classify_single_function}
\alias{classify_single_function}
\title{Classify a single function by heuristic signals}
\usage{
classify_single_function(fn_name, fml_names, rd_title)
}
\description{
Classify a single function by heuristic signals
}
\keyword{internal}
