% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{collect_calls}
\alias{collect_calls}
\title{Collect function calls by name from an AST recursively}
\usage{
collect_calls(expr, fn_name)
}
\description{
Collect function calls by name from an AST recursively
}
\keyword{internal}
