% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{collect_conditional_stops}
\alias{collect_conditional_stops}
\title{Collect stop()/warning() calls inside if-conditions referencing params}
\usage{
collect_conditional_stops(expr, param_names)
}
\description{
Collect stop()/warning() calls inside if-conditions referencing params
}
\keyword{internal}
