% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plugin.R
\name{collect_node_hints}
\alias{collect_node_hints}
\title{Collect computable_hint and skip_hint expressions from a node}
\usage{
collect_node_hints(node)
}
\description{
Collect computable_hint and skip_hint expressions from a node
}
\keyword{internal}
