% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plugin.R
\name{compute_node_order}
\alias{compute_node_order}
\title{Compute topological order of graph nodes via BFS}
\usage{
compute_node_order(graph)
}
\description{
Compute topological order of graph nodes via BFS
}
\keyword{internal}
