% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{ConstraintSet}
\alias{ConstraintSet}
\title{ConstraintSet}
\arguments{
\item{package}{Target package name (character).}

\item{func}{Target function name (character).}

\item{constraints}{A list of \link{Constraint} objects.}
}
\description{
A collection of constraints for a target package function.
Constraint IDs must be unique within the set.
}
