% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_schema.R
\name{ContextSchema}
\alias{ContextSchema}
\title{ContextSchema}
\arguments{
\item{variables}{A list of \link{ContextVariable} objects.}

\item{data_expectations}{Optional list of \link{DataExpectation} objects.}
}
\description{
The context schema defining variable scope and data expectations
for a plugin.
}
