% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{crossref_get}
\alias{crossref_get}
\title{Perform HTTP GET to CrossRef API}
\usage{
crossref_get(doi, mailto = NULL, timeout = 10)
}
\description{
Perform HTTP GET to CrossRef API
}
\keyword{internal}
