% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{detect_rd_deprecated}
\alias{detect_rd_deprecated}
\title{Detect deprecated parameters from Rd descriptions}
\usage{
detect_rd_deprecated(descriptions)
}
\description{
Detect deprecated parameters from Rd descriptions
}
\keyword{internal}
