% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridle_agent.R
\name{.load_manifest}
\alias{.load_manifest}
\title{Load manifest.yaml from plugin directory}
\usage{
.load_manifest(plugin_dir)
}
\description{
Load manifest.yaml from plugin directory
}
\keyword{internal}
