% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_engine.R
\name{engine_current_node}
\alias{engine_current_node}
\title{Get the current node ID from a GraphEngine}
\usage{
engine_current_node(engine)
}
\arguments{
\item{engine}{A \link{GraphEngine} object.}
}
\value{
Character string of the current node ID.
}
\description{
Get the current node ID from a GraphEngine
}
