% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_engine.R
\name{engine_visit_counts}
\alias{engine_visit_counts}
\title{Get visit counts from a GraphEngine}
\usage{
engine_visit_counts(engine)
}
\arguments{
\item{engine}{A \link{GraphEngine} object.}
}
\value{
Named integer vector of visit counts.
}
\description{
Get visit counts from a GraphEngine
}
