% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_engine.R
\name{evaluate_transitions}
\alias{evaluate_transitions}
\title{Evaluate all transition candidates for a node}
\usage{
evaluate_transitions(engine, node_id)
}
\arguments{
\item{engine}{A \link{GraphEngine} object.}

\item{node_id}{Node ID (character).}
}
\value{
A list of \link{TransitionCandidate} objects.
}
\description{
Evaluate all transition candidates for a node
}
