% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{exclude_s3_methods}
\alias{exclude_s3_methods}
\title{Exclude S3 methods from a list of exported names}
\usage{
exclude_s3_methods(exports, package)
}
\description{
Exclude S3 methods from a list of exported names
}
\keyword{internal}
