% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{extract_abstract_from_xml}
\alias{extract_abstract_from_xml}
\title{Extract abstract text from PubMed XML response}
\usage{
extract_abstract_from_xml(xml_text)
}
\description{
Extract abstract text from PubMed XML response
}
\keyword{internal}
