% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{extract_char_vector}
\alias{extract_char_vector}
\title{Extract character vector from a c() expression}
\usage{
extract_char_vector(expr)
}
\description{
Extract character vector from a c() expression
}
\keyword{internal}
