% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{extract_crossref_authors}
\alias{extract_crossref_authors}
\title{Extract author names from CrossRef author list}
\usage{
extract_crossref_authors(author_list)
}
\description{
Extract author names from CrossRef author list
}
\keyword{internal}
