% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{extract_rd_param_descriptions}
\alias{extract_rd_param_descriptions}
\title{Extract parameter descriptions from the \\arguments section}
\usage{
extract_rd_param_descriptions(rd)
}
\description{
Extract parameter descriptions from the \\arguments section
}
\keyword{internal}
