% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{extract_rd_references}
\alias{extract_rd_references}
\title{Extract references from the \\references section}
\usage{
extract_rd_references(rd)
}
\description{
Extract references from the \\references section
}
\keyword{internal}
