% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{extract_stop_message}
\alias{extract_stop_message}
\title{Extract message string from a stop() or warning() call}
\usage{
extract_stop_message(call_expr)
}
\description{
Extract message string from a stop() or warning() call
}
\keyword{internal}
