% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{extract_values_from_text}
\alias{extract_values_from_text}
\title{Extract enumerated values from a text string}
\usage{
extract_values_from_text(text)
}
\description{
Extract enumerated values from a text string
}
\keyword{internal}
