% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{fetch_pubmed_abstract}
\alias{fetch_pubmed_abstract}
\title{Fetch abstract from PubMed Entrez API (mockable wrapper)}
\usage{
fetch_pubmed_abstract(title, timeout = 10)
}
\description{
Fetch abstract from PubMed Entrez API (mockable wrapper)
}
\keyword{internal}
