% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_graph.R
\name{find_exit_transitions}
\alias{find_exit_transitions}
\title{Find post-exit root nodes in the function graph}
\usage{
find_exit_transitions(func_nodes_raw, entry_node)
}
\description{
Identifies function-graph nodes that have no incoming transitions
from other function-graph nodes and are not the entry_node. These
are the "post-exit" roots that the template's exit_point should
connect to.
}
\keyword{internal}
