% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{find_function_rd}
\alias{find_function_rd}
\title{Find the Rd object for a specific function by name or alias}
\usage{
find_function_rd(rd_db, func_name)
}
\description{
Find the Rd object for a specific function by name or alias
}
\keyword{internal}
