% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{find_match_arg_values}
\alias{find_match_arg_values}
\title{Find match.arg() calls and extract valid values}
\usage{
find_match_arg_values(fn_body, fmls, param_names)
}
\description{
Find match.arg() calls and extract valid values
}
\keyword{internal}
