% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{find_stop_constraints}
\alias{find_stop_constraints}
\title{Find stop()/warning() calls with parameter-related conditions}
\usage{
find_stop_constraints(fn_body, param_names, package, func)
}
\description{
Find stop()/warning() calls with parameter-related conditions
}
\keyword{internal}
