% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{get_namespace_exports}
\alias{get_namespace_exports}
\title{Get exported function names from a package namespace (mockable)}
\usage{
get_namespace_exports(package)
}
\description{
Get exported function names from a package namespace (mockable)
}
\keyword{internal}
