% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{get_package_namespace}
\alias{get_package_namespace}
\title{Get a package namespace (mockable wrapper)}
\usage{
get_package_namespace(package)
}
\description{
Get a package namespace (mockable wrapper)
}
\keyword{internal}
