% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{get_package_version}
\alias{get_package_version}
\title{Get installed package version (mockable wrapper)}
\usage{
get_package_version(package)
}
\description{
Get installed package version (mockable wrapper)
}
\keyword{internal}
