% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{get_rd_section}
\alias{get_rd_section}
\title{Get a specific section from an Rd object by tag}
\usage{
get_rd_section(rd, tag_name)
}
\description{
Get a specific section from an Rd object by tag
}
\keyword{internal}
