% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{get_s3_method_table}
\alias{get_s3_method_table}
\title{Get S3 methods table for a package (mockable)}
\usage{
get_s3_method_table(package)
}
\description{
Get S3 methods table for a package (mockable)
}
\keyword{internal}
