% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{is_conditional_default}
\alias{is_conditional_default}
\title{Check if an expression is a conditional default (ifelse/switch/if)}
\usage{
is_conditional_default(expr)
}
\description{
Check if an expression is a conditional default (ifelse/switch/if)
}
\keyword{internal}
