% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{is_deprecation_expr}
\alias{is_deprecation_expr}
\title{Detect lifecycle deprecation patterns in default expressions}
\usage{
is_deprecation_expr(expr)
}
\description{
Detect lifecycle deprecation patterns in default expressions
}
\keyword{internal}
