% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{KnowledgeStore}
\alias{KnowledgeStore}
\title{KnowledgeStore}
\arguments{
\item{topic}{The decision point this knowledge addresses (character).}

\item{target_parameter}{R package parameter name(s) (character vector).}

\item{package}{Target package name (character).}

\item{func}{Target function name (character).}

\item{entries}{A list of \link{KnowledgeEntry} objects.}
}
\description{
A collection of knowledge entries for a single topic, linked to
decision graph nodes via the \code{topic} field.
}
