% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_graph.R
\name{merge_template}
\alias{merge_template}
\title{Merge template nodes into a function-specific graph}
\usage{
merge_template(graph_raw, tmpl)
}
\description{
Merge template nodes into a function-specific graph
}
\keyword{internal}
