% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{parse_match_arg}
\alias{parse_match_arg}
\title{Parse a single match.arg() call to extract param and choices}
\usage{
parse_match_arg(call_expr, fmls, param_names)
}
\description{
Parse a single match.arg() call to extract param and choices
}
\keyword{internal}
