% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{pubmed_efetch_get}
\alias{pubmed_efetch_get}
\title{Perform PubMed efetch HTTP request}
\usage{
pubmed_efetch_get(pmid, timeout = 10)
}
\description{
Perform PubMed efetch HTTP request
}
\keyword{internal}
