% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{pubmed_fetch_abstract}
\alias{pubmed_fetch_abstract}
\title{Fetch abstract text from PubMed by PMID}
\usage{
pubmed_fetch_abstract(pmid, timeout = 10)
}
\description{
Fetch abstract text from PubMed by PMID
}
\keyword{internal}
