% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{pubmed_search_get}
\alias{pubmed_search_get}
\title{Perform PubMed search HTTP request}
\usage{
pubmed_search_get(title, timeout = 10)
}
\description{
Perform PubMed search HTTP request
}
\keyword{internal}
