% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_references.R
\name{rate_limit_sleep}
\alias{rate_limit_sleep}
\title{Wrapper around Sys.sleep for testability}
\usage{
rate_limit_sleep(seconds)
}
\description{
Wrapper around Sys.sleep for testability
}
\keyword{internal}
