% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{read_constraints}
\alias{read_constraints}
\title{Read Constraints from YAML}
\usage{
read_constraints(path)
}
\arguments{
\item{path}{Path to the YAML file.}
}
\value{
A \link{ConstraintSet} object.
}
\description{
Parses a constraints YAML file and returns a \link{ConstraintSet} object.
}
