% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{resolve_function}
\alias{resolve_function}
\title{Resolve a function from a package namespace}
\usage{
resolve_function(package, func)
}
\description{
Resolve a function from a package namespace
}
\keyword{internal}
