% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{safe_deparse}
\alias{safe_deparse}
\title{Safely deparse an expression to a single string}
\usage{
safe_deparse(expr)
}
\description{
Safely deparse an expression to a single string
}
\keyword{internal}
