% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{scan_package}
\alias{scan_package}
\title{Package Scanner}
\usage{
scan_package(package, func)
}
\arguments{
\item{package}{Package name (character).}

\item{func}{Function name (character).}
}
\value{
A \link{ScanResult} object.
}
\description{
Analyzes an R package function to produce a \link{ScanResult}. Layer 1 extracts
parameter names, default expressions, AST-parsed dependency graphs, and
automatic parameter classification (ADR-0004).

Entry point for the plugin generation scanner. Currently implements Layer 1
(formals analysis). Layers 2 and 3 will be added in subsequent Issues.
}
