% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_package.R
\name{update_deprecated_params}
\alias{update_deprecated_params}
\title{Update parameter classifications for deprecated params detected by Rd}
\usage{
update_deprecated_params(parameters, deprecated_params)
}
\description{
Update parameter classifications for deprecated params detected by Rd
}
\keyword{internal}
