% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_knowledge.R
\name{write_draft_files}
\alias{write_draft_files}
\title{Write draft YAML files to the output directory}
\usage{
write_draft_files(drafts, output_dir, package, func)
}
\description{
Write draft YAML files to the output directory
}
\keyword{internal}
